Revert env, clean up constants, and fetchSongs
