+++ .env
@@ -1,1 +1,1 @@
-SPOTIFY_KEY="BQBMZZosnvwfDL5NcaswvimxW3c-60FipV6VgwqLBQH4zwCyzrokb_su7xBqPe5i75jfDz8_ct6YWxD0VYXXYvE--kpf3UQQZxbVzqX0MsQA3_70VHiR38GOaGh2cgXX7tyUbssoGM0njwISNtQ02hDE2xmis_RQmCCHCLg"
+SPOTIFY_KEY=""
