+++ src/constants.ts
@@ -17,5 +17,128 @@ export const LOFI_MOOD = [
   0, 0, 0, 0, 0, 0
 ]
 
+export const attributesMap = {
+  "0": {
+    "attribute" : "danceability",
+    "ulimit" : 0.3,
+    "llimit" : 0,
+    "desc" : "No Dance"
+  },
+  "1" : {
+    "attribute" : "danceability",
+    "ulimit" : 0.5,
+    "llimit" : 0.3,
+    "desc" : "Slow Panced Dance"
+  },
+  "2" : {
+    "attribute" : "danceability",
+    "ulimit" : 0.7,
+    "llimit" : 0.5,
+    "desc" : "High Paced Dance"
+  },
+  "3" : {
+    "attribute" : "danceability",
+    "ulimit" : 0.7,
+    "llimit" : 1,
+    "desc" : "EDM"
+  },
+  "4" : {
+    "attribute" : "energy",
+    "ulimit" : 0.3,
+    "llimit" : 0,
+    "desc" : "Calm Energy"
+  },
+  "5" : {
+    "attribute" : "energy",
+    "ulimit" : 0.5,
+    "llimit" : 0.3,
+    "desc" : "Med-low Energy"
+  },
+  "6" : {
+    "attribute" : "energy",
+    "ulimit" : 0.5,
+    "llimit" : 0.7,
+    "desc" : "Med-high Energy"
+  },
+  "7" : {
+    "attribute" : "energy",
+    "ulimit" : 0.7,
+    "llimit" : 1,
+    "desc" : "High Energy"
+  },
+  "8" : {
+    "attribute" : "instrumentalness",
+    "ulimit" : 0.3,
+    "llimit" : 0,
+    "desc" : "Less Instrumental"
+  },
+  "9" : {
+    "attribute" : "instrumentalness",
+    "ulimit" : 0.5,
+    "llimit" : 0.3,
+    "desc" : "Medium Instrumental"
+  },
+  "10" : {
+    "attribute" : "instrumentalness",
+    "ulimit" : 0.5,
+    "llimit" : 0.7,
+    "desc" : "Medium-high Instrumental"
+    
+  },
+  "11" : {
+    "attribute" : "instrumentalness",
+    "ulimit" : 0.7,
+    "llimit" : 1,
+    "desc" : "Very Instrumental"
+  },
+  "12" : {
+    "attribute" : "speechiness",
+    "ulimit" : 0.3,
+    "llimit" : 0,
+    "desc" : "Maybe Instrumental"
+  },
+  "13" : {
+    "attribute" : "speechiness",
+    "ulimit" : 0.5,
+    "llimit" : 0.3,
+    "desc" : "Few Words"
+  },
+  "14" : {
+    "attribute" : "speechiness",
+    "ulimit" : 0.5,
+    "llimit" : 0.7,
+    "desc" : "Lots Of Words"
+  },
+  "15" : {
+    "attribute" : "speechiness",
+    "ulimit" : 0.7,
+    "llimit" : 1,
+    "desc" : "Rap"
+  },
+  "16" : {
+    "attribute" : "valence",
+    "ulimit" : -0.5,
+    "llimit" : -1,
+    "desc" : "Angry"
+  },
+  "17" : {
+    "attribute" : "valence",
+    "ulimit" : 0,
+    "llimit" : -0.5,
+    "desc" : "Depressed"
+  },
+  "18" : {
+    "attribute" : "valence",
+    "ulimit" : 0.5,
+    "llimit" : 0,
+    "desc" : "Sad"
+  },
+  "19" : {
+    "attribute" : "valence",
+    "ulimit" : 1,
+    "llimit" : 0.5,
+    "desc" : "Euphoric"
+  }
+}
 // pop https://open.spotify.com/playlist/37i9dQZF1DXarRysLJmuju?si=f09ee0c7a00a4cfb
 // kevin 4E5a8jO5lkzKJebOXeLUhC
